ores, properties, and one structure change modified description, name, authors, and licence

+++ src/main/java/com/github/cao/awa/trtr/heat/conductor/MetalBlockHeatConductor.java
@@ -21,7 +21,7 @@ public class MetalBlockHeatConductor extends HeatConductor {
         return super.getTemperature();
     }
 
-    public void setTemperature(int temperature) {
+    public void setTemperature(double temperature) {
         super.setTemperature(temperature);
     }
